Separate choices from basic UI into another class/file
